requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
django-rest-authtoken==2.1.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
pillow==10.3.0
PyJWT==1.7.1
pyotp==2.9.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
